// src/lib/api/generated/model/scheduleRuleKind.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Turnerito API
 * API del turnero SaaS multi-tenant
 * OpenAPI spec version: 1.0
 */

export type ScheduleRuleKind = typeof ScheduleRuleKind[keyof typeof ScheduleRuleKind];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ScheduleRuleKind = {
  work: 'work',
  break: 'break',
} as const;